+++ src/treecipe/src/FakerService/SnowfakeryFakerService/tests/mocks/SnowfakeryMockService.ts
@@ -1,0 +1,5 @@
+
+export class SnowfakeryMockService {
+
+
+}
